refactor: unify jscodeshift error handling
